+++ start/src/App.vue
@@ -7,26 +7,21 @@
     FIconClock
   } from '@fighting-design/fighting-icon'
 
-  const open2 = ref(false)
+  const isOpen = ref(false)
 </script>
 
 <template>
-  <f-sticky-card :open="open2">
-    <template #source>
-      <h1>1212121</h1>
-      <h1>1212121</h1>
-      <h1>1212121</h1>
-    </template>
+  <f-switch v-model="isOpen" size="middle" />
 
-    <h1>Hello World！</h1>
-    <h1>Hello World！</h1>
-    <h1>Hello World！</h1>
-    <h1>Hello World！</h1>
-    <h1>Hello World！</h1>
-    <h1>Hello World！</h1>
-    <h1>Hello World！</h1>
-    <h1>Hello World！</h1>
-  </f-sticky-card>
+  <f-collapse-animation :opened="isOpen">
+    <h1>Hello</h1>
+    <h1>你看到我了吗？</h1>
+
+    <f-button type="primary">主要按钮</f-button>
+    <f-button type="success">成功按钮</f-button>
+    <f-button type="danger">危险按钮</f-button>
+    <f-button type="warning">警告按钮</f-button>
+  </f-collapse-animation>
 
   <f-menu mode="inline">
     <f-submenu>