+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 中文 | [英文](https://github.com/FightingDesign/fighting-design/blob/master/CHANGELOG.en-US.md)
 
+- 修复 `f-pagination` 组件细节问题
+
 ## 0.9.0-alpha.5 (2022-11-25)
 
 **新功能**