+++ CHANGELOG.md
@@ -2,7 +2,17 @@
 
 中文 | [英文](https://github.com/FightingDesign/fighting-design/blob/master/CHANGELOG.en-US.md)
 
+**新功能**
+
+- 新增 `f-menu` 组件
+- 新增 `f-menu-item` 组件
+- 新增 `f-menu-item-group` 组件
+- 新增 `f-submenu` 组件
+
+**修复**
+
 - 修复 `f-pagination` 组件细节问题
+- 修复 `f-button` 组件样式细节
 
 ## 0.9.0-alpha.5 (2022-11-25)